Checker texture in separate module
